An unsupervised chatter detection method based on AE and DBSCAN clustering utilizing Internal CNC machine Signals
In the field of manufacturing, chatter is a common and undesirable phenomenon that can lead to reduced product quality and increased tool wear.
Detecting chatter in real-time is crucial to prevent these issues and improve overall machining efficiency. In this paper, we propose an innovative unsupervised chatter detection method using Autoencoders (AE) and Density-Based Spatial Clustering of Applications with Noise (DBSCAN) clustering algorithm, which leverages the internal signals of Computer Numerical Control (CNC) machines.